Renamed a few files and changed the imports
