+++ server/server.js
@@ -4,12 +4,13 @@ require("dotenv").config();
 //require express to create and configure our HTTP server
 const express = require("express");
 const app = express();
-const cors = require('cors')
-app.use(cors())
+const cors = require("cors");
+app.use(cors());
 //require mongoose to connect to our MongoDB database
 const mongoose = require("mongoose");
 
-const medicineRoutes = require("./routes/medicinesRoutes");
+const medicineRoutes = require("./routes/medicineRoutes");
+// const appointmentRoutes = require("./models/appointmentRoutes");
 
 // middleware
 app.use(express.json());
@@ -21,6 +22,7 @@ app.use((req, res, next) => {
 
 // routes
 app.use("/api/medicines", medicineRoutes);
+// app.use("/api/appointments", appointmentRoutes);
 
 // connect to db
 mongoose
